% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_12month_periods.R
\name{add_12month_periods}
\alias{add_12month_periods}
\title{Adds a column that classifies observations into 12-month periods}
\usage{
add_12month_periods(df, start, end)
}
\arguments{
\item{df}{A dataframe containing columns \code{ano} and \code{mes}.}

\item{start}{Character string indicating beggining of the first 12-month period: YYYY-MM}

\item{end}{Character string indicating the end of the last 12-month period. Recommended format: YYYY-MM}
}
\value{
Original dataframe with an additional \code{periodo} variable.
}
\description{
Adds a column that classifies observations into 12-month periods
}
\seealso{
\code{\link{read_caged_rds}}
}
