% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geometry.R
\name{add_geometry_municipios}
\alias{add_geometry_municipios}
\title{Adds a geometry column to dataframe with municipality-level data}
\usage{
add_geometry_municipios(df)
}
\arguments{
\item{df}{A dataframe containing data from cities belonging to the state of
Sao Paulo. The dataframe must have a numeric column named \emph{Codmun7}
with cities' IBGE 7-digit code.}
}
\value{
A dataframe containing the original data and an additional
  \emph{geometry} column with municipality-level geospatial data.
}
\description{
Joins provided dataframe with another containing municipality-level
\code{MULTIPOLYGON} geospatial data. The key used to perform the join is the
numeric column \emph{Codmun7}, which must contain cities' 7-digit code, as
used by IBGE.
}
\examples{
municipios_sp \%>\%
  add_geometry_municipios()
}
