% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geometry.R
\name{add_geometry_regioes_gov}
\alias{add_geometry_regioes_gov}
\title{Adds a geometry column to dataframe with region-level data}
\usage{
add_geometry_regioes_gov(df)
}
\arguments{
\item{df}{A dataframe containing data from Sao Paulo's regions
(\emph{Regiões de Governo}). It must contains a column named
\emph{regiao_governo}.}
}
\value{
A dataframe containing the original data and an additional
  \emph{geometry} column with region-level geospatial data.
}
\description{
Joins provided dataframe with another containing region-level
\code{POLYGON} geospatial data. The key used to perform the join is the
character column \emph{regiao_governo}.
}
\examples{
municipios_sp \%>\%
  count(regiao_governo) \%>\%
  add_geometry_municipios()
}
