% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_monthly.R
\name{barplot_monthly}
\alias{barplot_monthly}
\title{Barplot showing the monthly evolution of job creation}
\usage{
barplot_monthly(
  df,
  inicio = NA,
  fim = NA,
  incluir_ano_x = FALSE,
  labs_title = "Empregos no Estado de SP - Evolução mensal",
  labs_subtitle = "Contribuição de cada setor",
  labs_caption = "Fonte: Elaboração própria a partir de dados do CAGED.",
  labs_x = "Mês",
  labs_y = "Postos de trabalho (milhares)"
)
}
\arguments{
\item{df}{A dataframe with CAGED micro-data.}

\item{inicio}{Optional string indicating the starting month, if one does not
to use all available data. Recommended format: YYYY-MM.}

\item{fim}{Optional string indicating the final month, if one does not to use
all available data. Recommended format: YYYY-MM.}

\item{incluir_ano_x}{\code{logical}. If \code{TRUE}, x-axis labels include
the year of each month. Recommended when plotting months of more than a
single year. Defaults to \code{FALSE}.}

\item{labs_title}{Optional character string with plot title.}

\item{labs_subtitle}{Optional character string with plot subtitle.}

\item{labs_caption}{Optional character string with plot caption.}

\item{labs_x}{Optional character string with x-axis title.}

\item{labs_y}{Optional character string with y-axis title.}
}
\value{
A \pkg{ggplot2} plot.
}
\description{
Builds a barplot depicting the evolution of job creation over time, by month.
Each bar indicates the jobs created/destroyed in a month. Colors represent
economic sectors (IBGE's "big-five"). Labels joined by dashed lines show
monthly net job creation.
}
