% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blues_full}
\alias{blues_full}
\title{Set of blue HEX codes}
\format{A character vector.}
\source{
Created using \pkg{RColorBrewer}.
}
\usage{
blues_full
}
\description{
A character vector containing HEX codes of 9 intensities of blue,
from lightest to darkest.
}
\keyword{datasets}
