% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Returns string in upper-case without accents, tildes and cedillas}
\usage{
clean_text(string)
}
\arguments{
\item{string}{A character string}
}
\value{
A clean, upper-case version of provided \code{string}
}
\description{
\code{clean_text} converts the provided string to upper-case
and removes accents, tildes and cedillas.
}
\examples{
clean_text('São Paulo')
clean_text('Mogi-Mirim')
clean_text(c('Carapicuíba', 'Tibiriçá', 'SERTÃOZINHO'))
}
