% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnae_classes}
\alias{cnae_classes}
\title{Lookup table of CNAE classes codes and names}
\format{A dataframe with 673 rows and 2 variables:
\describe{
\item{codigo}{CNAE numeric code}
\item{classe_cnae}{CNAE class}
}}
\source{
\url{https://concla.ibge.gov.br/agencia-sala-de-imprensa/classificacoes/download-concla/8265-download.html}
}
\usage{
cnae_classes
}
\description{
A dataframe with numeric codes and names of economic activities,
according to the National Classification of Economic Activities
 - CNAE.
}
\keyword{datasets}
