% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_job_creation.R
\name{compute_job_creation}
\alias{compute_job_creation}
\title{Computes job creation, job loss and net job creation}
\usage{
compute_job_creation(df, ...)
}
\arguments{
\item{df}{A dataframe with CAGED micro-data. Must have a \code{movimento}
column.}

\item{...}{Additional variables for aggretated summaries.}
}
\value{
A dataframe.
}
\description{
Computes job creation, job loss and net job creation, given a dataframe with
CAGED micro-data. The input dataframe must have a column named
\code{movimento}, whose values take values 1 (job creation) or -1 (job
destruction).
}
\details{
Allows for aggregated summaries by passing grouping variables in \code{...}
}
