% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{criar_tabela_municipio}
\alias{criar_tabela_municipio}
\title{Creates summary table for a given city}
\usage{
criar_tabela_municipio(
  df_ano_atual,
  df_ano_anterior,
  mes_atual,
  municipio,
  colunas = NA,
  csv_output = NA
)
}
\arguments{
\item{df_ano_atual}{A dataframe.}

\item{df_ano_anterior}{A dataframe.}

\item{mes_atual}{An integer between 1 and 12.}

\item{municipio}{A character string with the city name.}

\item{colunas}{A character vector with the column names of the output
dataframe.}

\item{csv_output}{A character string with the csv output path. Optional.}
}
\value{
A dataframe.
}
\description{
Creates summary table for a given city
}
