% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{criar_tabela_regiao_adm}
\alias{criar_tabela_regiao_adm}
\title{Create a summary table for a given Admnistrative Region}
\usage{
criar_tabela_regiao_adm(
  df_ano_atual,
  df_ano_anterior,
  mes_atual,
  regiao_adm = "Ribeirão Preto",
  colunas = NA,
  csv_output = NA
)
}
\arguments{
\item{df_ano_atual}{A dataframe.}

\item{df_ano_anterior}{A dataframe.}

\item{mes_atual}{An integer between 1 and 12.}

\item{regiao_adm}{A character string with the region name.}

\item{colunas}{A character vector with the column names of the output
dataframe.}

\item{csv_output}{A character string with the csv output path. Optional.}
}
\value{
A dataframe.
}
\description{
Create a summary table for a given Admnistrative Region
}
