% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{custom_theme}
\alias{custom_theme}
\title{Custom \code{ggplot2} theme settings}
\usage{
custom_theme()
}
\value{
A \code{ggplot2} \code{theme} object
}
\description{
Custom \code{ggplot2} theme settings
}
\examples{
ggplot(data = mtcars) +
geom_point(aes(x = mpg, y = cyl)) +
custom_theme()
}
