% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{decode_variables}
\alias{decode_variables}
\title{Decodes numeric-encoded categoric variables from CAGED micro-data}
\usage{
decode_variables(df, vars = "all")
}
\arguments{
\item{df}{A dataframe with encoded categoric variables.}

\item{vars}{A character vector with supported variable names or "all".}
}
\value{
A dataframe, with additional columns for the decoded variables.
}
\description{
Creates new variables with proper labels, given a dataframe with
numeric-encoded categoric variables.
}
\details{
Supported variables: \itemize{ \item
\code{codigo_cnae} \item \code{codigo_grau_instrucao} \item
\code{codigo_raca_cor} \item \code{codigo_faixa_tamanho_estab} \item
\code{codigo_sexo} }
}
