% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{drop_codes}
\alias{drop_codes}
\title{Drops columns containing numeric-encoded categoric variables}
\usage{
drop_codes(df)
}
\arguments{
\item{df}{A dataframe containing the numeric-encoded variables listed above.}
}
\value{
The same dataframe, but without the listed variables.
}
\description{
Removes the following variables from a dataframe: \itemize{ \item
\code{codigo_uf} \item \code{codigo_raca_cor} \item \code{codigo_sexo}
\item \code{codigo_grau_instrucao} \item \code{codigo_sexo} }
}
\examples{
\dontrun{
read_caged() \%>\%
  decode_variables('all') \%>\%
  drop_codes()
}
}
