% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faixa_tamanho_estab}
\alias{faixa_tamanho_estab}
\title{Lookup table of codes and labels of company sizes}
\format{A dataframe with 10 rows and 2 variables:
\describe{
\item{codigo}{Numeric code, ranging from 1 (up to 4 employees) to 9 (more than 1000 employees) (-1 for \code{NA})}
\item{faixa_tamanho_estab}{Company size bin}
}}
\source{
\url{ftp://ftp.mtps.gov.br/pdet/microdados/CAGED/CAGEDEST_layout_Atualizado.xls}
}
\usage{
faixa_tamanho_estab
}
\description{
In CAGED microdata, companies are binned according to the number of
people they employ. This dataframe contains codes and labels for
each bin.
}
\keyword{datasets}
