% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cnae_highlights.R
\name{find_cnae_highlights}
\alias{find_cnae_highlights}
\title{Returns a dataframe with job creation data on highlighted CNAE sectors}
\usage{
find_cnae_highlights(
  df,
  positivos = TRUE,
  n_destaques = 5,
  filtrar_setor = FALSE
)
}
\arguments{
\item{df}{A dataframe containing summarized job creation data. It must have
columns \code{setor} and \code{saldo}.}

\item{positivos}{\code{logical}. If \code{FALSE}, function returns
worst-performing observations.}

\item{n_destaques}{Number of observations to include as highlights.}

\item{filtrar_setor}{\code{FALSE} or one of the 'big-five' IBGE sectors. If a
sector is provided, the function returns only highlights for that specific
sector.}
}
\value{
A dataframe with job creation data on CNAE sectors with best (or
  worst) performance.
}
\description{
Identifies best- or worst-performing CNAE sectors with respect to job creation
and returns a dataframe with selected data.
}
\examples{
\dontrun{
read_caged('CAGED_2018_10.txt') \%>\%
  prepare_caged() \%>\%
  filter(municipio_clean == 'RIBEIRAO PRETO') \%>\%
  compute_job_creation(setor, classe_cnae, .drop = FALSE) \%>\%
  find_cnae_highlights(setor == 'Serviços')
}
}
