% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_number.R
\name{format_number}
\alias{format_number}
\title{Helper function to adjust number formatting}
\usage{
format_number(x, ...)
}
\arguments{
\item{x}{A numeric or character vector.}

\item{...}{Arguments passed to \code{formatC}.}
}
\value{
A character object.
}
\description{
Helper function to adjust number formatting
}
\examples{
format_number(1000000)
format_number(c('15.72', '1,000.42'))
}
