% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_breaks.R
\name{get_breaks}
\alias{get_breaks}
\title{Returns custom break points for plots}
\usage{
get_breaks(x, n = 9)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{Number of groups defined by the break points}
}
\value{
A numeric vector with the break points
}
\description{
Returns a numeric vector with custom break points. Break points are defined so as to isolate
the highest positive value and lowest negative value, include zero, and split the remaining
observations into groups of similar sizes.
}
\examples{
get_breaks(rnorm(n = 100, mean = 0, sd = 100))
get_breaks(c(-100, seq(-80, -10, 10), 0, seq(10, 80, 10) 100))
}
