% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{get_ibge_sector}
\alias{get_ibge_sector}
\title{Returns IBGE's 5 'big sectors' from numeric vector of IBGE' subsectors}
\usage{
get_ibge_sector(x = codigo_subsetor_ibge, return_factor = TRUE)
}
\arguments{
\item{x}{A numeric vector, with integers ranging from 1 to 25.}

\item{return_factor}{Logical. If \code{TRUE}, returns a factor. Else, returns
a character vector.}
}
\value{
A factor or character vector with the corresponding sector.
}
\description{
Relationship between economic subsectors and sectors: \itemize{ \item
\strong{[1, 14]}: Indústria \item \strong{15}: Construção Civil \item
\strong{[16, 17]}: Comércio \item \strong{[18, 24]}: Serviços \item
\strong{25}: Agricultura }
}
\examples{
get_ibge_sector(x = 1:25, return_factor = FALSE)
}
