% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\name{get_palette}
\alias{get_palette}
\title{Returns a custom 'RdBu' color palette}
\usage{
get_palette(map_breaks)
}
\arguments{
\item{map_breaks}{A numeric vector}
}
\value{
A character vector with HEX codes
}
\description{
Builds a custom 'RdBu'color palette based on a
vector of breaks.
}
