% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grau_instrucao}
\alias{grau_instrucao}
\title{Lookup table of education levels codes and labels}
\format{A dataframe with 12 rows and 2 variables:
\describe{
\item{codigo}{Numeric code, ranging from 1 (illiterate) to 11 (PhD) (-1 for \code{NA})}
\item{grau_instrucao}{Education level, in Portuguese}
}}
\source{
\url{ftp://ftp.mtps.gov.br/pdet/microdados/CAGED/CAGEDEST_layout_Atualizado.xls}
}
\usage{
grau_instrucao
}
\description{
A dataframe with codes and labels of education levels used
in CAGED microdata.
}
\keyword{datasets}
