% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{join_municipios_sp}
\alias{join_municipios_sp}
\title{Adds municipality-level data to a dataframe containing CAGED microdata}
\usage{
join_municipios_sp(df)
}
\arguments{
\item{df}{A dataframe. It must contain a numeric column
\code{codigo_municipio}, with the 6-digit city code.}
}
\value{
A dataframe with additional columns.
}
\description{
Includes new variables with data regarding cities from the State of Sao
Paulo. New data comes from dataset \code{municipios_sp}, and includes:
\itemize{
\item \code{Codmun7}
\item \code{municipio}
\item \code{municipio_clean}
\item \code{pop2019}
\item \code{regiao_admnisitrativa}
\item \code{regiao_governo}
\item \code{regiao_metropolitana}
\item \code{aglomeracao_urbana}
}
}
\details{
\strong{This function supports cities from the State of Sao Paulo only.
Thus, it is recommended to filter the input dataframe beforehand.}
}
\examples{
\dontrun{
read_caged() \%>\%
  join_dados_municipios()
}
}
