% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_region_sequential.R
\name{map_region_sequential}
\alias{map_region_sequential}
\title{Plots a sequential-scale map of Sao Paulo's Government Regions}
\usage{
map_region_sequential(
  df,
  var_plot,
  map_title = str_c("Regiões Gov. de SP - ", var_plot, " - Acumulado 12 meses"),
  map_breaks = NA,
  map_palette = "Blues"
)
}
\arguments{
\item{df}{A dataframe containing one line per Government Region, and a
variable named \code{regiao_governo}.}

\item{var_plot}{\code{character} with the name of the variable to be plotted.}

\item{map_title}{Optional character string with map title.}

\item{map_breaks}{Optional numeric vector with breaks.}

\item{map_palette}{Character string with the name of a sequential color
palette or character vector with color HEX codes.}
}
\value{
A \pkg{tmap} plot.
}
\description{
Wrapper around \pkg{tmap} calls to draw a map of Government Regions of the
State of Sao Paulo, using a sequential color scheme.
}
