% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meses}
\alias{meses}
\title{Lookup table of month abbreviations in Portuguese}
\format{A dataframe with 12 rows and 2 variables:
\describe{
\item{numero}{Month number, from 1 to 12}
\item{abrev}{Month abbreviation, in Portuguese}
}}
\usage{
meses
}
\description{
A dataframe with month numbers and abbreviations in Portuguese.
}
\keyword{datasets}
