% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipios_sp}
\alias{municipios_sp}
\title{Dataframe with information regarding cities belongin to the State of Sao Paulo}
\format{A dataframe with 645 rows and 9 variables:
\describe{
\item{Codmun7}{City 7-digit code}
\item{codigo}{City 6-digit code}
\item{municipio}{City name}
\item{municipio_clean}{City name, all uppercase, without accents, cedillas and tildes}
\item{pop}{City population in 2019}
\item{regiao_administrativa}{City's Admnistrative Region}
\item{regiao_governo}{City's Government Region}
\item{regiao_metropolitana}{City's Metropolitan Area, if applicable}
\item{aglomeracao_urbana}{City's urban agglomeration, if applicable}
}}
\source{
IBGE
}
\usage{
municipios_sp
}
\description{
A dataset with information regarding the 645 Brazilian cities from
 the State of Sao Paulo. Useful in joins, to find cities' population
 and regions, using city codes or names as keys.
}
\keyword{datasets}
