% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygons_municipios_sp}
\alias{polygons_municipios_sp}
\title{\pkg{sf} dataframe with geospatial data on cities from the State of Sao Paulo}
\format{A \pkg{sf} dataframe with 645 features (rows), 2 fields (variables),
 and geometry:
\describe{
\item{NM_MUNICIP}{City name, all uppercase, with accents, tildes and cedillas}
\item{Codmun7}{City 7-digit code}
\item{geometry}{\code{MULTIPOLYGON} geometry with coordinates of city's boundaries}
}}
\usage{
polygons_municipios_sp
}
\description{
A dataset of boundaries for each of the 645 Brazilian cities belonging
to the State of Sao Paulo.
}
\keyword{datasets}
