% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygons_regioes_gov_sp}
\alias{polygons_regioes_gov_sp}
\title{\pkg{sf} dataframe with geospatial data on regions from the State of Sao Paulo}
\format{A \pkg{sf} dataframe with 43 features (rows), 1 field (variable),
 and geometry:
\describe{
\item{NM_MUNICIP}{Region name, title case, with accents, tildes and cedillas}
\item{geometry}{Coordinates of region boundaries}
}}
\usage{
polygons_regioes_gov_sp
}
\description{
A dataset of boundaries for each of the 43 Government Regions from the
State of Sao Paulo.
}
\keyword{datasets}
