% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_caged}
\alias{prepare_caged}
\title{Prepares CAGED micro-data to data analysis focused on the State of Sao Paulo}
\usage{
prepare_caged(df)
}
\arguments{
\item{df}{A dataframe similar to the output of \code{read_caged()}.}
}
\value{
A dataframe with additional data.
}
\description{
Filters CAGED microdata to observations belonging to the State of Sao Paulo,
inludes additional variables containing municipality-level information and
decodes numeric-encoded categoric variables.
}
\examples{
\dontrun{
read_caged('CAGEDEST_102018.txt') \%>\%
prepare_caged()
}
}
