% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raca_cor}
\alias{raca_cor}
\title{Lookup table of race/color codes and labels}
\format{A dataframe with 7 rows and 2 variables:
\describe{
\item{codigo}{Numeric code, ranging from 1 to 9 (-1 for \code{NA})}
\item{raca_cor}{Person's race/color}
}}
\source{
\url{ftp://ftp.mtps.gov.br/pdet/microdados/CAGED/CAGEDEST_layout_Atualizado.xls}
}
\usage{
raca_cor
}
\description{
A dataframe with numeric codes and labels of races and colors,
as used in CAGED microdata.
}
\keyword{datasets}
