% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_caged}
\alias{read_caged}
\title{Reads CAGED data from txt file}
\usage{
read_caged(txt_file, pretty = TRUE, ...)
}
\arguments{
\item{txt_file}{Character string indicating the txt file path.}

\item{pretty}{Logical. If \code{TRUE}, uses well formatted column names
instead of those from the txt file.}

\item{...}{Additional arguments to \code{read_delim}.}
}
\value{
A dataframe with data from the txt file.
}
\description{
Uses \code{read_delim()} to read a flat file containing monthly data from
CAGED, as provided by the Brazilian Ministry of Labor.
}
\examples{
\dontrun{
caged_out2018 <- read_caged('CAGEDEST_102018.txt')
}
}
