% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_caged_rds}
\alias{read_caged_rds}
\title{Reads one or multiple .rds files with preprocessed CAGED data}
\usage{
read_caged_rds(
  data_folder,
  start,
  end,
  periodos_12meses = TRUE,
  nested = FALSE,
  stringAsFactors = TRUE
)
}
\arguments{
\item{data_folder}{A character string indicating the directory where the rds
files are located.}

\item{start}{A character string indicating the month and year of the
\emph{first} month. Recommended format: YYYY-MM.}

\item{end}{A character string indicating the month and year of the
\emph{last} month. Recommended format: YYYY-MM.}

\item{periodos_12meses}{Logical. Can be set to \code{TRUE} only if multiples
of 12 rds files are read. If \code{TRUE}, the resulting dataframe contains
a column indicating the 12-month period each observation belongs to.}

\item{nested}{Logical. Defaults to \code{FALSE}. If \code{TRUE}, the function
returns a nested dataframe with a single row per rds file.}

\item{stringAsFactors}{Logical.}
}
\value{
A dataframe.
}
\description{
Loops through rds files containing dataframes with CAGED preprocessed data
from a given month, reads them, and organizes data in a single dataframe.
}
\details{
\strong{Note:} rds files must be located inside the directory passed in the
\code{data_folder} argument, and must be named in the following way:
\emph{CAGED_YYYY_MM.rds}.
}
\examples{
\dontrun{
caged2012 <-
read_caged_rds(data_folder = 'caged_clean',
               start = '2012-01', end = '2012-12')
}
}
