% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reds_full}
\alias{reds_full}
\title{Set of red HEX codes}
\format{A character vector.}
\source{
Created using \pkg{RColorBrewer}.
}
\usage{
reds_full
}
\description{
A character vector containing HEX codes of 11 intensities of red,
from darkest to ligthest. Useful when adjusting plots colors manually.
}
\keyword{datasets}
