% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subsetor_ibge}
\alias{subsetor_ibge}
\title{Lookup table of IBGE's subsectors of economic activities}
\format{A dataframe with 21 rows and 2 variables:
\describe{
\item{codigo}{Subsector numeric code}
\item{subsetor_ibge}{Subsector name}
}}
\source{
IBGE
}
\usage{
subsetor_ibge
}
\description{
A dataframe with numeric codes and labels of IBGE's
 economic subsectors.
}
\keyword{datasets}
