% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tipo_deficiencia}
\alias{tipo_deficiencia}
\title{Lookup table of disability types}
\format{A dataframe with 8 rows and 2 variables:
\describe{
\item{codigo}{Numeric code}
\item{tipo_deficiencia}{Type of disability}
}}
\source{
\url{ftp://ftp.mtps.gov.br/pdet/microdados/CAGED/CAGEDEST_layout_Atualizado.xls}
}
\usage{
tipo_deficiencia
}
\description{
A dataframe with numeric codes and labels of
 disability types as used in CAGED microdata.
}
\keyword{datasets}
