% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tipo_mov_desag}
\alias{tipo_mov_desag}
\title{Lookup table of admission/termination categories}
\format{A dataframe with 15 rows and 2 variables:
\describe{
\item{codigo}{Numeric code}
\item{tipo_mov_desag}{Admission or termination type}
}}
\source{
\url{ftp://ftp.mtps.gov.br/pdet/microdados/CAGED/CAGEDEST_layout_Atualizado.xls}
}
\usage{
tipo_mov_desag
}
\description{
A dataframe with numeric codes and labels of
 admission/termination categories
}
\keyword{datasets}
