% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uf}
\alias{uf}
\title{Lookup table of Brazilian Federative Units (UFs) codes and abbreviations}
\format{A dataframe with 27 rows and 2 variables:
\describe{
\item{codigo}{Numeric code}
\item{sigla_uf}{Abbreviation}
}}
\usage{
uf
}
\description{
A dataframe with numeric codes and abbreviations of Brazilian States and
Federal District.
}
\keyword{datasets}
